server/requirements.txt
elevenlabs
python-dotenv
flask
requests